pyproject.toml
[tool.poetry]
name = "actions"
version = "0.1.0"
description = "airbyte actions"
authors = ["cpdeethree <conor@airbyte.io>", "alafanechere <augustin@airbyte.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dagger-io = "^0.5.1"
python-dotenv = "^0.21.0"
asyncclick = "^8.1.3.4"
google-cloud-run = "^0.5.0"
pre-commit = "^2.21.0"
mypy = "^1.1.1"
asyncer = "^0.0.2"
pygithub = "^1.58.1"
gitpython = "^3.1.31"

[tool.mypy]
python_version = "3.11"
strict = "true"
explicit_package_bases = "true" # we need this to make mypy work with script imports (importing sibling script from _pipelines)
namespace_packages = "true"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
